Refactor Prototype1b section reader
Separate the reader from the DataStorage class
Extract types and move types to separate namespace for Prototype1b
Add try/catch across DWG Prototype1b reader and DXF ACIS reader
