+++ src/ACadSharp/IO/DWG/DwgStreamReaders/DwgPrototype1bReader.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using ACadSharp.Prototype1b;
+using ACadSharp.Prototype1b.Segments;
+using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
@@ -8,10 +10,12 @@ using System.Text;
 namespace ACadSharp.IO.DWG.DwgStreamReaders
 {
     internal class DwgPrototype1bReader : DwgSectionIO
-    {
-        public override string SectionName => DwgSectionDefinition.AcDsPrototype;
+	{
+		public static readonly uint[] TYPE_SIZES = new uint[] { 0, 0, 2, 1, 2, 4, 8, 1, 2, 4, 8, 4, 8, 0, 0, 0 };
+		private readonly IDwgStreamReader _reader;
+		private readonly DataStorage _storage = new DataStorage();
 
-        private readonly IDwgStreamReader _reader;
+		public override string SectionName => DwgSectionDefinition.AcDsPrototype;
 
         public DwgPrototype1bReader(ACadVersion version, IDwgStreamReader reader) : base(version)
         {
@@ -20,114 +24,44 @@ namespace ACadSharp.IO.DWG.DwgStreamReaders
 
         public DataStorage Read()
         {
-            // Everything appears to be aligned to 16 byte boundaries
-            return new DataStorage(this._reader);
-        }
-    }
+			try
+			{
+				this._storage.FileHeader = this.readFileHeader();
 
-    public class DataStoragePointers
-    {
-        // Indices for reading Segments, Schemas and Data
-        public SegmentIndex SegmentIndex { get; set; }
-        public DataIndex DataIndex { get; set; }
-        public SchemaIndex SchemaIndex { get; set; }
+				// Indices for reading values
+				this._storage.IndexPointers = new DataStoragePointers();
+				this._storage.IndexPointers.SegmentIndex = this.readSegmentIndex();
+				this._storage.IndexPointers.DataIndex = this.readDataIndex();
+				this._storage.IndexPointers.SchemaIndex = this.readSchemaIndex();
+				this._storage.IndexPointers.FreeSpace = this.readFreeSpace();		// Index to empty spaces (padding or so) within the file
 
-        // Empty areas inside the file which are filled with zeros for the most part (could be commented out
-        // sections as sometimes it contains invalid segment data that should be ignored)
-        public FreeSpace FreeSpace { get; set; }
-    }
+				// Probably the file state before the last save
+				this._storage.PreviousSave = this.readPreviousSave();
 
-    public class DataStorage
-    {
-        private static readonly uint[] TYPE_SIZES = new uint[] { 0, 0, 2, 1, 2, 4, 8, 1, 2, 4, 8, 4, 8, 0, 0, 0 };
+				// Schema data lookup
+				this._storage.SchemaSearch = this.readSchemaSearch();
+				foreach (SchemaSearchEntry search in this._storage.SchemaSearch.Entries) {
+					search.SchemaName = this._storage.IndexPointers.SchemaIndex.SchemaNames[search.SchemaNameIndex];
+				}
 
-        private readonly IDwgStreamReader _reader;
+				this._storage.SchemaFields = [];
+				this._storage.DataFields = [];
+				this._storage.Blobs = [];
 
-        internal DataStorage(IDwgStreamReader reader)
-        {
-            this._reader = reader;
-            this.FileHeader = this.readFileHeader();
+				// From this point on it should be possible to read storage entries sequentially
+				// The issue is that sometimes there seemingly are empty padding sections (which should be
+				// referenced by the FreeSpace entry) that are missing from the FreeSpace definition or are larger
+				// than specified in the FreeSpace definition.
 
-            // Indices for reading values
-            this.IndexPointers = new DataStoragePointers();
-            this.IndexPointers.SegmentIndex = this.readSegmentIndex();
-            this.IndexPointers.DataIndex = this.readDataIndex();
-            this.IndexPointers.SchemaIndex = this.readSchemaIndex();
-            this.IndexPointers.FreeSpace = this.readFreeSpace();		// Index to empty spaces (padding or so) within the file
+				this.ReadSegments();
+			}
+			catch (Exception ex)
+			{
+				this.notify("An error occurred while reading the Prototype1b", NotificationType.Error, ex);
+			}
 
-            // Probably the file state before the last save
-            this.PreviousSave = this.readPreviousSave();
-
-            // Schema data lookup
-            this.SchemaSearch = this.readSchemaSearch();
-            foreach (SchemaSearchEntry search in this.SchemaSearch.Entries) {
-                search.SchemaName = this.IndexPointers.SchemaIndex.SchemaNames[search.SchemaNameIndex];
-            }
-
-            this.SchemaFields = [];
-            this.DataFields = [];
-            this.Blobs = [];
-
-            // From this point on it should be possible to read storage entries sequentially
-            // The issue is that sometimes there seemingly are empty padding sections (which should be
-            // referenced by the FreeSpace entry) that are missing from the FreeSpace definition or are larger
-            // than specified in the FreeSpace definition.
-
-            this.ReadSegments();
-        }
-
-        public FileHeader FileHeader { get; set; }
-
-        // Information about file state at last revision
-        public PreviousSave PreviousSave { get; set; }
-
-        public DataStoragePointers IndexPointers { get; set; }
-
-        // All registered Schemas
-        public List<SchemaData> SchemaFields { get; set; }
-
-        // The search object to get the data entries associated with schema entries
-        public SchemaSearch SchemaSearch { get; set; }
-        public List<DataField> DataFields { get; set; }
-        public List<Blob01> Blobs { get; set; }
-
-        public Schema GetSchemaByName(string name)
-        {
-            foreach (SchemaData data in this.SchemaFields) {
-                foreach (Schema schema in data.Values) {
-                    if (schema.Name == name) {
-                        return schema;
-                    }
-                }
-            }
-            return null;
-        }
-
-        public List<DataEntry> GetSchemaData(Schema schema)
-        {
-            if (schema == null) return [];
-
-            foreach (SchemaSearchEntry search in this.SchemaSearch.Entries) {
-                if (search.SchemaNameIndex == schema.Index) {
-                    List<DataEntry> data = [];
-                    Dictionary<ulong, DataEntry> allDataEntries = this.DataFields
-                        .SelectMany(x => x.Entries)
-                        .Where(x => x.Header.SchemaIndex == schema.Index)
-                        .GroupBy(x => x.Header.Handle)
-                        .ToDictionary(x => x.Key, x => x.First());	// TODO: Find out how to correctly handle the case where duplicate handles exist, but have different values
-
-                    foreach (SearchEntryObject[] entries in search.IdEntryObjects) {
-                        foreach (SearchEntryObject entry in entries) {
-                            if (entry.Indices.Length == 0) continue;	// TODO: Find out what this index really means (maybe index inside the data field entry where the item is?) and why there can be multiple arrays and why some have none at all
-                            data.Add(allDataEntries[entry.Handle]);
-                        }
-                    }
-                    return data;
-                }
-            }
-
-            return [];
-        }
+			return this._storage;
+		}
 
         internal IPrototype1bSegment ReadSegmentAt(ulong offset)
         {
@@ -140,39 +74,39 @@ namespace ACadSharp.IO.DWG.DwgStreamReaders
 
         private SegmentIndex readSegmentIndex()
         {
-            return (SegmentIndex)this.ReadSegmentAt((ulong)this.FileHeader.SegmentIndexOffset);
+            return (SegmentIndex)this.ReadSegmentAt((ulong)this._storage.FileHeader.SegmentIndexOffset);
         }
 
         private DataIndex readDataIndex()
         {
-            return (DataIndex)this.ReadSegmentAt(this.IndexPointers.SegmentIndex.Pointers[this.FileHeader.DataIndexSegmentIndex].Offset);
+            return (DataIndex)this.ReadSegmentAt(this._storage.IndexPointers.SegmentIndex.Pointers[this._storage.FileHeader.DataIndexSegmentIndex].Offset);
         }
 
         private SchemaSearch readSchemaSearch()
         {
-            return (SchemaSearch)this.ReadSegmentAt(this.IndexPointers.SegmentIndex.Pointers[this.FileHeader.SearchSegmentIndex].Offset);
+            return (SchemaSearch)this.ReadSegmentAt(this._storage.IndexPointers.SegmentIndex.Pointers[this._storage.FileHeader.SearchSegmentIndex].Offset);
         }
 
         private FreeSpace readFreeSpace()
         {
-            if (this.FileHeader.FreeSpaceSegmentIndex == 0) return null;
-            return (FreeSpace)this.ReadSegmentAt(this.IndexPointers.SegmentIndex.Pointers[this.FileHeader.FreeSpaceSegmentIndex].Offset);
+            if (this._storage.FileHeader.FreeSpaceSegmentIndex == 0) return null;
+            return (FreeSpace)this.ReadSegmentAt(this._storage.IndexPointers.SegmentIndex.Pointers[this._storage.FileHeader.FreeSpaceSegmentIndex].Offset);
         }
 
         private PreviousSave readPreviousSave()
         {
-            if (this.FileHeader.PreviousSaveIndex == 0) return null;
-            return (PreviousSave)this.ReadSegmentAt(this.IndexPointers.SegmentIndex.Pointers[this.FileHeader.PreviousSaveIndex].Offset);
+            if (this._storage.FileHeader.PreviousSaveIndex == 0) return null;
+            return (PreviousSave)this.ReadSegmentAt(this._storage.IndexPointers.SegmentIndex.Pointers[this._storage.FileHeader.PreviousSaveIndex].Offset);
         }
 
         private SchemaIndex readSchemaIndex()
         {
-            return (SchemaIndex)this.ReadSegmentAt(this.IndexPointers.SegmentIndex.Pointers[this.FileHeader.SchemaIndexSegmentIndex].Offset);
+            return (SchemaIndex)this.ReadSegmentAt(this._storage.IndexPointers.SegmentIndex.Pointers[this._storage.FileHeader.SchemaIndexSegmentIndex].Offset);
         }
 
         internal void ReadSegments()
         {
-            foreach (KeyValuePair<int, SegmentIndexEntry> entry in this.IndexPointers.SegmentIndex.Pointers) {
+            foreach (KeyValuePair<int, SegmentIndexEntry> entry in this._storage.IndexPointers.SegmentIndex.Pointers) {
                 if (entry.Value.Size == 0) continue;
                 this._reader.Position = (long) entry.Value.Offset;
 
@@ -190,13 +124,13 @@ namespace ACadSharp.IO.DWG.DwgStreamReaders
 
                     // Add entries which are types that can appear multiple times to a collection of them
                     case SchemaData schdat:
-                        this.SchemaFields.Add(schdat);
+                        this._storage.SchemaFields.Add(schdat);
                         break;
                     case DataField data:
-                        this.DataFields.Add(data);
+                        this._storage.DataFields.Add(data);
                         break;
                     case Blob01 blob:
-                        this.Blobs.Add(blob);
+                        this._storage.Blobs.Add(blob);
                         break;
                     default:
                         break;
@@ -348,7 +282,7 @@ namespace ACadSharp.IO.DWG.DwgStreamReaders
                 Header = header,
                 Pointers = []
             };
-            for (int j = 0; j < this.FileHeader.SegmentIndexEntryCount; j++) {
+            for (int j = 0; j < this._storage.FileHeader.SegmentIndexEntryCount; j++) {
                 index.Pointers[j] = new SegmentIndexEntry {
                     Offset = this._reader.ReadRawULong(),
                     Size = this._reader.ReadUInt()
@@ -471,7 +405,7 @@ namespace ACadSharp.IO.DWG.DwgStreamReaders
             // Read headers
             List<DataHeader> headers = [];
             uint headerEndPosition = (uint)this._reader.Position;
-            if (!this.IndexPointers.DataIndex.Entries.TryGetValue(header.SegmentIndex, out DataIndexEntry entry)) {
+            if (!this._storage.IndexPointers.DataIndex.Entries.TryGetValue(header.SegmentIndex, out DataIndexEntry entry)) {
                 return dataField;
             }
 
@@ -600,7 +534,7 @@ namespace ACadSharp.IO.DWG.DwgStreamReaders
         {
             // Read unknown schema properties
             List<SchemaUnknownProperty> unknownProps = [];
-            foreach (SchemaPropertyPointer pointer in this.IndexPointers.SchemaIndex.SchemaUnknownPropertyPointer) {
+            foreach (SchemaPropertyPointer pointer in this._storage.IndexPointers.SchemaIndex.SchemaUnknownPropertyPointer) {
                 if (pointer.SchemaIndex != header.SegmentIndex) continue;
                 unknownProps.Add(new SchemaUnknownProperty {
                     DataSize = this._reader.ReadUInt(),
@@ -610,11 +544,11 @@ namespace ACadSharp.IO.DWG.DwgStreamReaders
 
             // Read schemas
             List<Schema> schemaValues = [];
-            foreach (SchemaPropertyPointer pointer in this.IndexPointers.SchemaIndex.SchemaPointer) {
+            foreach (SchemaPropertyPointer pointer in this._storage.IndexPointers.SchemaIndex.SchemaPointer) {
                 if (pointer.SchemaIndex != header.SegmentIndex) continue;
                 Schema schema = this.readSchema();
                 schema.Index = pointer.Index;
-                schema.Name = this.IndexPointers.SchemaIndex.SchemaNames[pointer.Index];
+                schema.Name = this._storage.IndexPointers.SchemaIndex.SchemaNames[pointer.Index];
                 schemaValues.Add(schema);
             }
 
@@ -650,7 +584,7 @@ namespace ACadSharp.IO.DWG.DwgStreamReaders
             FreeSpace space = new() {
                 Header = header,
                 Unknown = this._reader.ReadRawULong(),
-                FreeSpaces = new FreeSpaceArea[this.FileHeader.FreeSpaceEntryCount]
+                FreeSpaces = new FreeSpaceArea[this._storage.FileHeader.FreeSpaceEntryCount]
             };
             for (int i = 0; i < space.FreeSpaces.Length; i++) {		// TODO: Sometimes when there is an additional freespace definition, it might not match the `FreeSpaceEntryCount` count
                 ulong position = this._reader.ReadRawULong();
@@ -729,235 +663,5 @@ namespace ACadSharp.IO.DWG.DwgStreamReaders
 
             return search;
         }
-    }
-
-    public class FileHeader
-    {
-        public uint FileSignature { get; set; }
-        public int FileHeaderSize { get; set; }
-        public int UnknownFlag { get; set; }			// 0; 1
-        public int Unknown1 { get; set; }				// 8; 2
-        public int Version { get; set; }
-        public int Unknown2 { get; set; }				// 0
-        public int DataStorageRevision { get; set; }
-        public int SegmentIndexOffset { get; set; }				// segidx --> this._reader.Position of beginning
-        public int SegmentIndexUnknown { get; set; }	// 0
-        public int SegmentIndexEntryCount { get; set; }			// lastId / count of segidx entries
-        public int SchemaIndexSegmentIndex { get; set; }		// schidx
-        public int DataIndexSegmentIndex { get; set; }			// datidx
-        public int SearchSegmentIndex { get; set; }				// search
-        public int PreviousSaveIndex { get; set; }				// prvsav
-        public int FileSize { get; set; }
-        public int Unknown3 { get; set; }				// 0
-        public int FreeSpaceSegmentIndex { get; set; }			// freesp
-        public int FreeSpaceEntryCount { get; set; }
-        public int Unknown5 { get; set; }				// 0
-
-        public byte[] UnknownRemaining { get; set; }	// Zeros
-    }
-
-    public interface IPrototype1bSegment { 
-        public SegmentHeader Header { get; set; }
-    }
-
-    public class PreviousSave : IPrototype1bSegment
-    {
-        public SegmentHeader Header { get; set; }
-        public FileHeader FileHeader { get; set; }
-    }
-
-    public class DataField : IPrototype1bSegment
-    {
-        public SegmentHeader Header { get; set; }
-        public List<DataEntry> Entries { get; set; }
-    }
-
-    public class DataEntry
-    {
-        public DataHeader Header { get; set; }
-        public DataValue Value { get; set; }
-    }
-
-    public class DataValue
-    {
-        public uint DataSize { get; set; }
-        public byte[] Data { get; set; }
-        public DataBlobReference BlobReference { get; set; }
-    }
-
-    public class DataBlobReference
-    {
-        public ulong TotalDataSize { get; set; }
-        public uint PageCount { get; set; }
-        public uint RecordSize { get; set; }    // DataBlobReference size
-        public uint PageSize { get; set; }
-        public uint LastPageSize { get; set; }
-        public uint Unknown1 { get; set; }
-        public uint Unknown2 { get; set; }
-        public List<(uint, uint)> SegmentPointers { get; set; }		// Pointer (segment index, size) to blob01 segment containing the data
-    }
-
-    public class DataIndex : IPrototype1bSegment
-    {
-        public SegmentHeader Header { get; set; }
-        public int Unknown1 { get; set; }
-        public Dictionary<uint, DataIndexEntry> Entries { get; set; }
-    }
-
-    public class DataIndexEntry
-    {
-        public uint DataSegmentIndex { get; set; }
-        public List<DataIndexEntryPointer> Pointers { get; set; }
-    }
-
-    public class DataIndexEntryPointer
-    {
-        public uint Offset { get; set; }
-        public uint SchemaIndex { get; set; }
-    }
-
-    public class Blob01 : IPrototype1bSegment
-    {
-        public SegmentHeader Header { get; set; }
-        public ulong TotalDataSize { get; set; }
-        public ulong PageStartOffset { get; set; }
-        public uint PageIndex { get; set; }
-        public uint PageCount { get; set; }
-        public ulong PageDataSize { get; set; }
-        public byte[] Data { get; set; }
-    }
-
-    public class SchemaData : IPrototype1bSegment
-    {
-        public SegmentHeader Header { get; set; }
-        public List<SchemaUnknownProperty> SchemaUnknownProperties { get; set; }
-        public List<Schema> Values { get; set; }
-    }
-
-    public class SchemaUnknownProperty
-    {
-        public uint DataSize { get; set; }
-        public uint UnknownFlags { get; set; }
-    }
-
-    public class Schema
-    {
-        public uint Index { get; set; }
-        public string Name { get; set; }
-        public ulong[] Indices { get; set; }
-        // TODO: Find out what PropertyDescriptors are
-        public SchemaProperty[] Properties { get; set; }
-    }
-
-    public class SchemaProperty
-    {
-        public uint? Type { get; set; }
-        public uint PropertyFlags { get; set; }
-        public uint NameIndex { get; set; }
-        public string Name { get; set; }
-        public uint TypeSize { get; set; }
-        public uint Unknown1 { get; set; }
-        public uint Unknown2 { get; set; }
-        public uint PropertyValueCount { get; set; }
-        public byte[,] Values { get; set; }
-    }
-
-    public class SchemaIndex : IPrototype1bSegment
-    {
-        public SegmentHeader Header { get; set; }
-        public uint Unknown1 { get; set; }			// 0; 1431655765		--> 0; 0x5555 0x5555
-        public long UnknownMagic { get; set; }		// 0x0af10c
-        public SchemaPropertyPointer[] SchemaPointer { get; set; }
-        public SchemaPropertyPointer[] UnknownPropertyEntryPointer { get; set; }
-        public SchemaPropertyPointer[] SchemaUnknownPropertyPointer { get; set; }
-        public uint UnknownIndex1 { get; set; }		// 5; 4
-        public string[] SchemaNames { get; set; }
-    }
-
-    public class SchemaPropertyPointer
-    {
-        public uint SchemaIndex { get; set; }
-        public uint Index { get; set; }
-        public uint Offset { get; set; }
-    }
-
-    public class SchemaSearch : IPrototype1bSegment
-    {
-        public SegmentHeader Header { get; set; }
-        public List<SchemaSearchEntry> Entries { get; set; }
-    }
-
-    public class FreeSpace : IPrototype1bSegment
-    {
-        public SegmentHeader Header { get; set; }
-        public ulong Unknown { get; set; }
-        public FreeSpaceArea[] FreeSpaces { get; set; }
-    }
-
-    public class FreeSpaceArea
-    {
-        public ulong Position { get; set; }
-        public uint Size { get; set; }
-    }
-
-    public class SegmentIndex : IPrototype1bSegment
-    {
-        public SegmentHeader Header { get; set; }
-        public Dictionary<int, SegmentIndexEntry> Pointers { get; set; }
-    }
-
-    public class SegmentIndexEntry
-    {
-        public ulong Offset { get; set; }
-        public uint Size { get; set; }
-    }
-
-    public class SchemaSearchEntry
-    {
-        public uint SchemaNameIndex { get; set; }
-        public ulong[] SortedIndices { get; set; }
-        public uint Unknown1 { get; set; }
-        public SearchEntryObject[][] IdEntryObjects { get; set; }
-
-        // Resolved values
-        public string SchemaName { get; set; }
-    }
-
-    public class SearchEntryObject
-    {
-        public ulong Handle { get; set; }
-        public ulong[] Indices { get; set; }
-    }
-
-    public class SegmentHeader
-    {
-        /// <summary>
-        /// The size of the segment header. The <see cref="Name"/> is always 6 characters
-        /// long and the <see cref="AlignmentBytes"/> is always 8 bytes long.
-        /// </summary>
-        public const uint SIZE = 48;
-
-        public short Signature { get; set; }
-        public string Name { get; set; }
-        public uint SegmentIndex { get; set; }
-        public int IsBlob { get; set; }
-        public uint SegmentSize { get; set; }
-        public int Unknown1 { get; set; }
-        public int DataStorageRevision { get; set; }
-        public int Unknown2 { get; set; }
-        public int SystemDataAlignmentOffset { get; set; }
-        public int ObjectDataAlignmentOffset { get; set; }
-        public byte[] AlignmentBytes { get; set; }
-    }
-
-    public class DataHeader
-    {
-        public uint EntrySize { get; set; }
-        public uint Unknown1 { get; set; }
-        public ulong Handle { get; set; }
-        public uint DataOffset { get; set; }
-
-        // Resolved Values
-        public uint SchemaIndex { get; set; }
     }
 }
