+++ src/ACadSharp/Prototype1b/DataEntry.cs
@@ -1,0 +1,8 @@
+﻿namespace ACadSharp.Prototype1b
+{
+	public class DataEntry
+    {
+        public DataHeader Header { get; set; }
+        public DataValue Value { get; set; }
+    }
+}
