+++ src/ACadSharp/Prototype1b/DataIndexEntry.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace ACadSharp.Prototype1b
+{
+	public class DataIndexEntry
+    {
+        public uint DataSegmentIndex { get; set; }
+        public List<DataIndexEntryPointer> Pointers { get; set; }
+    }
+}
