+++ src/ACadSharp/Prototype1b/DataIndexEntryPointer.cs
@@ -1,0 +1,8 @@
+﻿namespace ACadSharp.Prototype1b
+{
+	public class DataIndexEntryPointer
+    {
+        public uint Offset { get; set; }
+        public uint SchemaIndex { get; set; }
+    }
+}
