+++ src/ACadSharp/Prototype1b/DataValue.cs
@@ -1,0 +1,9 @@
+﻿namespace ACadSharp.Prototype1b
+{
+	public class DataValue
+    {
+        public uint DataSize { get; set; }
+        public byte[] Data { get; set; }
+        public DataBlobReference BlobReference { get; set; }
+    }
+}
