+++ src/ACadSharp/Prototype1b/FreeSpaceArea.cs
@@ -1,0 +1,8 @@
+﻿namespace ACadSharp.Prototype1b
+{
+	public class FreeSpaceArea
+    {
+        public ulong Position { get; set; }
+        public uint Size { get; set; }
+    }
+}
