+++ src/ACadSharp/Prototype1b/Schema.cs
@@ -1,0 +1,11 @@
+﻿namespace ACadSharp.Prototype1b
+{
+	public class Schema
+    {
+        public uint Index { get; set; }
+        public string Name { get; set; }
+        public ulong[] Indices { get; set; }
+        // TODO: Find out what PropertyDescriptors are
+        public SchemaProperty[] Properties { get; set; }
+    }
+}
