+++ src/ACadSharp/Prototype1b/SchemaPropertyPointer.cs
@@ -1,0 +1,9 @@
+﻿namespace ACadSharp.Prototype1b
+{
+	public class SchemaPropertyPointer
+    {
+        public uint SchemaIndex { get; set; }
+        public uint Index { get; set; }
+        public uint Offset { get; set; }
+    }
+}
