+++ src/ACadSharp/Prototype1b/SchemaUnknownProperty.cs
@@ -1,0 +1,8 @@
+﻿namespace ACadSharp.Prototype1b
+{
+	public class SchemaUnknownProperty
+    {
+        public uint DataSize { get; set; }
+        public uint UnknownFlags { get; set; }
+    }
+}
