+++ src/ACadSharp/Prototype1b/SearchEntryObject.cs
@@ -1,0 +1,8 @@
+﻿namespace ACadSharp.Prototype1b
+{
+	public class SearchEntryObject
+    {
+        public ulong Handle { get; set; }
+        public ulong[] Indices { get; set; }
+    }
+}
