+++ src/ACadSharp/Prototype1b/SegmentIndexEntry.cs
@@ -1,0 +1,8 @@
+﻿namespace ACadSharp.Prototype1b
+{
+	public class SegmentIndexEntry
+    {
+        public ulong Offset { get; set; }
+        public uint Size { get; set; }
+    }
+}
