+++ src/ACadSharp/Prototype1b/Segments/DataField.cs
@@ -1,0 +1,11 @@
+﻿using ACadSharp.Prototype1b.Segments;
+using System.Collections.Generic;
+
+namespace ACadSharp.Prototype1b
+{
+	public class DataField : IPrototype1bSegment
+    {
+        public SegmentHeader Header { get; set; }
+        public List<DataEntry> Entries { get; set; }
+    }
+}
