+++ src/ACadSharp/Prototype1b/Segments/IPrototype1bSegment.cs
@@ -1,0 +1,6 @@
+﻿namespace ACadSharp.Prototype1b.Segments
+{
+	public interface IPrototype1bSegment { 
+        public SegmentHeader Header { get; set; }
+    }
+}
