+++ src/ACadSharp/Prototype1b/Segments/PreviousSave.cs
@@ -1,0 +1,10 @@
+﻿using ACadSharp.Prototype1b.Segments;
+
+namespace ACadSharp.Prototype1b
+{
+	public class PreviousSave : IPrototype1bSegment
+    {
+        public SegmentHeader Header { get; set; }
+        public FileHeader FileHeader { get; set; }
+    }
+}
